backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
pandas>=2.0,<3.0
scikit-learn>=1.4,<2.0
joblib>=1.3,<2.0
